olfati_django/requirements.txt
celery
colorlog
Django==5.2.0
django-debug-toolbar
djangorestframework
djangorestframework_simplejwt
drf-nested-routers
drf-spectacular
httpx
jsonschema
jsonschema-specifications
kombu
pillow
prompt_toolkit
psycopg2-binary
PyJWT
python-dateutil
python-decouple
PyYAML
whitenoise
django_celery_results
kavenegar
firebase-admin
drf-nested-routers